requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
click==8.1.7
Django==5.1
django-solo==2.3.0
exceptiongroup==1.2.2
fastapi==0.99.0
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.8
pillow==10.4.0
pydantic==1.10.18
pydantic_core==2.20.1
python-telegram-bot==20.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
